refactor: rename all providers to services
All providers are now deprecated. They will be deleted to the next major
release.
