+++ spec/commons/services/business/OrderLBS.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import { EntityOptions, PaginatedEntitiesOptions } from '../../../../src/interfaces'
 import { PaginatedListResult } from '../../../../src/modules/iris-module/interfaces'
-import { BusinessValidatorProvider, ErrorProvider } from '../../../../src/modules/iris-module/providers'
+import { BusinessValidatorService, ErrorService } from '../../../../src/modules/iris-module/services'
 import { OrderBE } from '../../objects/business/be/OrderBE'
 import { OrderState } from '../../objects/business/be/OrderState'
 import { OrderFilterQuery } from '../../objects/filter/OrderFilterQuery'
@@ -12,7 +12,7 @@ import { AmountCalculator } from './AmountCalculator'
 @Injectable()
 export class OrderLBS {
 
-  constructor(private readonly businessValidatorProvider: BusinessValidatorProvider, private readonly orderDAO: OrderDAO, private readonly amountCalculator: AmountCalculator, private readonly errorProvider: ErrorProvider) {
+  constructor(private readonly businessValidatorService: BusinessValidatorService, private readonly orderDAO: OrderDAO, private readonly amountCalculator: AmountCalculator, private readonly errorService: ErrorService) {
   }
 
   public async findAll(query?: PaginatedEntitiesOptions, filters?: OrderFilterQuery): Promise<OrderBE[]> {
@@ -30,7 +30,7 @@ export class OrderLBS {
   public async findById(id: number, query?: EntityOptions): Promise<OrderBE> {
     const order = await this.orderDAO.findById(id, query)
     if (!order) {
-      throw this.errorProvider.createEntityNotFoundBusinessException('orders', id)
+      throw this.errorService.createEntityNotFoundBusinessException('orders', id)
     }
     return order
   }
@@ -38,7 +38,7 @@ export class OrderLBS {
   public async createOrder(order: OrderBE, query?: EntityOptions): Promise<OrderBE> {
     order.state = OrderState.SAVED
     this.amountCalculator.calculateOrderAmount(order)
-    order = this.businessValidatorProvider.validate(order)
+    order = this.businessValidatorService.validate(order)
     return this.orderDAO.save(order, query)
   }
 
@@ -49,7 +49,7 @@ export class OrderLBS {
   public async updateOrderState(orderId: number, orderState: OrderState): Promise<OrderBE> {
     const order = await this.orderDAO.findById(orderId)
     if (!order) {
-      throw this.errorProvider.createEntityNotFoundBusinessException('order', orderId)
+      throw this.errorService.createEntityNotFoundBusinessException('order', orderId)
     }
     order.state = orderState
     return this.orderDAO.save(order)
