+++ spec/commons/services/data/OrderDAO.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
-import { ErrorProvider, LoggerProvider } from '../../../../src/modules/iris-module/providers'
+import { ErrorService, LoggerService } from '../../../../src/modules/iris-module/services'
 import { IrisDAO } from '../../../../src/services/data'
 import { OrderBE } from '../../objects/business/be/OrderBE'
 import { OrderFilterQuery } from '../../objects/filter/OrderFilterQuery'
 
 @Injectable()
 export class OrderDAO extends IrisDAO<OrderBE, OrderFilterQuery> {
-  constructor(@InjectRepository(OrderBE) orderRepository: Repository<OrderBE>, errorProvider: ErrorProvider, loggerProvider: LoggerProvider) {
+  constructor(@InjectRepository(OrderBE) orderRepository: Repository<OrderBE>, errorProvider: ErrorService, loggerProvider: LoggerService) {
     super(orderRepository, errorProvider, loggerProvider)
   }
 }
