+++ src/modules/iris-module/index.ts
@@ -3,5 +3,6 @@ export * from './decorators'
 export * from './interceptors'
 export * from './interfaces'
 export * from './providers'
+export * from './services'
 export * from './iris.context'
 export * from './iris.module'
