+++ src/modules/iris-module/middlewares/logging.middleware.ts
@@ -1,16 +1,16 @@
 import { Injectable, NestMiddleware } from '@nestjs/common'
 import URI from 'urijs'
-import { LoggerProvider } from '../providers'
+import { LoggerService } from '../services'
 
 @Injectable()
 export class LoggerMiddleware implements NestMiddleware {
 
-  constructor(private readonly logger: LoggerProvider) {
+  constructor(private readonly loggerService: LoggerService) {
   }
 
   public use(req: any, res: any, next: () => void): any {
-    this.logger.info(`verb=${req.method},uri=${new URI(req.url).pathname()},queryParams=${JSON.stringify(req.query)},ip=${req.ip}`)
+    this.loggerService.info(`verb=${req.method},uri=${new URI(req.url).pathname()},queryParams=${JSON.stringify(req.query)},ip=${req.ip}`)
     next()
-    this.logger.info(`statusCode=${res.statusCode}`)
+    this.loggerService.info(`statusCode=${res.statusCode}`)
   }
 }
