+++ src/modules/iris-module/providers/default-authentication.provider.ts
@@ -1,11 +1,9 @@
 import { Injectable } from '@nestjs/common'
-import { Request } from 'express'
-import { AuthenticatedUser, AuthenticationService } from '../interfaces'
+import { DefaultAuthenticationService } from '../services'
 
+/**
+ * @deprecated use DefaultAuthenticationService instead.
+ */
 @Injectable()
-export class DefaultAuthenticationProvider implements AuthenticationService {
-  public async getAuthenticatedUser(request: Request): Promise<AuthenticatedUser | undefined> {
-    return undefined
-  }
-
+export class DefaultAuthenticationProvider extends DefaultAuthenticationService {
 }
