+++ src/modules/iris-module/providers/default-authorization.provider.ts
@@ -1,16 +1,9 @@
 import { Injectable } from '@nestjs/common'
-import { Request } from 'express'
-import { AuthorizationService } from '../interfaces'
-import { LoggerProvider } from './logger.provider'
+import { DefaultAuthorizationService } from '../services'
 
+/**
+ * @deprecated use DefaultAuthorizationService instead.
+ */
 @Injectable()
-export class DefaultAuthorizationProvider implements AuthorizationService {
-  constructor(private readonly logger: LoggerProvider) {
-  }
-
-  public async validateAuthorization(request: Request, ...functions: string[]): Promise<boolean> {
-    this.logger.warn('You should implement your own AuthorizationService')
-    return false
-  }
-
+export class DefaultAuthorizationProvider extends DefaultAuthorizationService {
 }
