+++ src/modules/iris-module/providers/message.provider.ts
@@ -1,25 +1,10 @@
-import { Inject, Injectable } from '@nestjs/common'
-import { IRIS_CONFIG_OPTIONS } from '../../../constants'
-import { MessageFactory } from '../commons'
-import { IrisConfigOptions } from '../../config-module/config-holder'
+import { Injectable } from '@nestjs/common'
+import { MessageService } from '../services'
 
-export let messageSource: MessageProvider
-
+/**
+ * @deprecated use MessageService instead.
+ */
 @Injectable()
-export class MessageProvider {
-  private messageFactory: MessageFactory
-
-  constructor(@Inject(IRIS_CONFIG_OPTIONS) private irisConfigOptions: IrisConfigOptions) {
-    this.messageFactory = new MessageFactory({ resources: this.irisConfigOptions.messagesSources })
-    messageSource = this
-  }
-
-  public has(key: string): boolean {
-    return this.messageFactory.has(key)
-  }
-
-  public get(key: string, datas?: object): string {
-    return this.messageFactory.get(key, datas) || `missing key ${key} in properties`
-  }
+export class MessageProvider extends MessageService {
 
 }
