--- src/modules/iris-module/providers/request-holder.provider.ts
@@ -1,7 +1,0 @@
-import { Injectable, Scope } from '@nestjs/common'
-
-@Injectable({ scope: Scope.REQUEST })
-export class RequestHolder {
-  public traceId: string
-  public spanId: string
-}
