+++ src/modules/iris-module/services/default-authentication.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common'
+import { Request } from 'express'
+import { AuthenticatedUser, AuthenticationService } from '../interfaces'
+
+@Injectable()
+export class DefaultAuthenticationService implements AuthenticationService {
+  public async getAuthenticatedUser(request: Request): Promise<AuthenticatedUser | undefined> {
+    return undefined
+  }
+
+}
