+++ src/services/index.ts
@@ -1,2 +1,1 @@
-export * from './business'
 export * from './data'
